refactor: 테스트 컨텍스트 하나만 로드하도록 리팩토링
