+++ src/test/java/com/pet/domains/area/controller/CityControllerTest.java
@@ -16,23 +16,15 @@ import static org.springframework.restdocs.payload.PayloadDocumentation.response
 import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.get;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import com.pet.domains.docs.BaseDocumentationTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
-import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 
-@WebMvcTest(value = CityController.class)
-@AutoConfigureRestDocs
 @DisplayName("시도/시군구 컨트롤러 테스트")
-class CityControllerTest {
-
-    @Autowired
-    private MockMvc mockMvc;
+class CityControllerTest extends BaseDocumentationTest {
 
     @Test
     @DisplayName("시도/시군구 조회 성공 테스트")
