+++ src/test/java/com/pet/domains/post/controller/MissingPostControllerTest.java
@@ -23,8 +23,8 @@ import static org.springframework.restdocs.request.RequestDocumentation.paramete
 import static org.springframework.restdocs.request.RequestDocumentation.pathParameters;
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import com.fasterxml.jackson.databind.ObjectMapper;
 import com.pet.common.jwt.JwtMockToken;
+import com.pet.domains.docs.BaseDocumentationTest;
 import com.pet.domains.post.domain.SexType;
 import com.pet.domains.post.domain.Status;
 import com.pet.domains.post.dto.request.MissingPostCreateParam;
@@ -33,24 +33,12 @@ import java.time.LocalDate;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
-import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 
-@WebMvcTest(MissingPostController.class)
-@AutoConfigureRestDocs
 @DisplayName("실종/보호 게시물 컨트롤러 테스트")
-class MissingPostControllerTest {
-
-    @Autowired
-    private MockMvc mockMvc;
-
-    @Autowired
-    private ObjectMapper objectMapper;
+class MissingPostControllerTest extends BaseDocumentationTest {
 
     @DisplayName("실종/보호 게시물 등록 테스트")
     @Test
@@ -315,7 +303,7 @@ class MissingPostControllerTest {
         // then
         resultActions
             .andExpect(status().isCreated())
-            .andDo(document("create-missing-post-bookmark}",
+            .andDo(document("create-missing-post-bookmark",
                 preprocessRequest(prettyPrint()),
                 preprocessResponse(prettyPrint()),
                 requestHeaders(
@@ -338,7 +326,7 @@ class MissingPostControllerTest {
         // then
         resultActions
             .andExpect(status().isNoContent())
-            .andDo(document("delete-missing-post-bookmark}",
+            .andDo(document("delete-missing-post-bookmark",
                 preprocessRequest(prettyPrint()),
                 preprocessResponse(prettyPrint()),
                 requestHeaders(
@@ -347,6 +335,49 @@ class MissingPostControllerTest {
                 pathParameters(
                     parameterWithName("postId").description("실종/보호 게시글 id")
                 ))
+            );
+    }
+
+    @Test
+    @DisplayName("특정 실종게시글의 댓글 리스트 조회 테스트")
+    void getShelterPostsTest() throws Exception {
+        // given
+        // when
+        ResultActions resultActions = mockMvc.perform(get("/api/v1/missing-posts/{postId}/comments", 1L)
+            .accept(MediaType.APPLICATION_JSON_VALUE)
+            .header(HttpHeaders.AUTHORIZATION, JwtMockToken.MOCK_TOKEN));
+
+        // then
+        resultActions
+            .andDo(print())
+            .andExpect(status().isOk())
+            .andDo(document("get-missing-posts-comments",
+                preprocessRequest(prettyPrint()),
+                preprocessResponse(prettyPrint()),
+                pathParameters(
+                    parameterWithName("postId").description("실종 게시글 아이디")
+                ),
+                requestHeaders(
+                    headerWithName(HttpHeaders.AUTHORIZATION).description("jwt token"),
+                    headerWithName(HttpHeaders.ACCEPT).description(MediaType.APPLICATION_JSON_VALUE)
+                ),
+                responseHeaders(
+                    headerWithName(HttpHeaders.CONTENT_TYPE).description(MediaType.APPLICATION_JSON_VALUE)
+                ),
+                responseFields(
+                    fieldWithPath("data").type(OBJECT).description("응답 데이터"),
+                    fieldWithPath("data.comments").type(ARRAY).description("댓글 목록"),
+                    fieldWithPath("data.comments[].id").type(NUMBER).description("댓글 아이디"),
+                    fieldWithPath("data.comments[].content").type(STRING).description("댓글 내용"),
+                    fieldWithPath("data.comments[].createdAt").type(STRING).description("댓글 작성날짜"),
+                    fieldWithPath("data.comments[].user").type(OBJECT).description("댓글 작성자"),
+                    fieldWithPath("data.comments[].user.id").type(NUMBER).description("작성자 아이디"),
+                    fieldWithPath("data.comments[].user.nickname").type(STRING).description("작성자 닉네임"),
+                    fieldWithPath("data.comments[].user.image").type(STRING).description("작성자 프로필 사진"),
+                    fieldWithPath("data.totalElements").type(NUMBER).description("전체 결과 수"),
+                    fieldWithPath("data.last").type(BOOLEAN).description("마지막 페이지 여부"),
+                    fieldWithPath("data.size").type(NUMBER).description("페이지당 요청 수"),
+                    fieldWithPath("serverDateTime").type(STRING).description("서버 응답 시간")))
             );
     }
