+++ src/test/java/com/pet/domains/post/controller/ShelterPostControllerTest.java
@@ -22,23 +22,15 @@ import static org.springframework.restdocs.request.RequestDocumentation.pathPara
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
 import com.pet.common.jwt.JwtMockToken;
+import com.pet.domains.docs.BaseDocumentationTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.restdocs.AutoConfigureRestDocs;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.MediaType;
-import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.test.web.servlet.ResultActions;
 
-@WebMvcTest(value = ShelterPostController.class)
-@AutoConfigureRestDocs
-@DisplayName("보호소 동물 게시글 컨트롤러 docs 테스트")
-class ShelterPostControllerTest {
-
-    @Autowired
-    private MockMvc mockMvc;
+@DisplayName("보호소 동물 게시글 컨트롤러 테스트")
+class ShelterPostControllerTest extends BaseDocumentationTest {
 
     @Test
     @DisplayName("보호소 게시글 리스트 조회 테스트")
